pyproject.toml
[project]
name = "multidoc-rag"
version = "0.1.0"
description = "Streamlined multi-document RAG QA app using LangChain + Groq + Chroma"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
  "typer[all]>=0.9",
  "python-dotenv>=1.0",
  "langchain-community>=0.1",
  "langchain-chroma>=0.1",
  "langchain-text-splitters>=0.1",
  "langchain-huggingface>=0.1",
  "langchain-groq>=0.1",
  "unstructured[local-inference,pdf]>=0.7",
  "nltk>=3.8",
  "pytest>=7.0",
  "rich>=13.0"
]

[project.scripts]
mdrag = "multidoc_rag.app:main"

[build-system]
requires = ["setuptools>=61.0","wheel"]
build-backend = "setuptools.build_meta"